+++ CURLHandleSource/CURLSocketRegistration.h
@@ -30,7 +30,7 @@
  * @param multi The multi that this object is working with.
  */
 
-- (void)updateSourcesForSocket:(int)socket mode:(NSInteger)mode multi:(CURLMultiHandle*)multi;
+- (void)updateSourcesForSocket:(int)socket mode:(int)mode multi:(CURLMultiHandle*)multi;
 
 /**
  Indicates whether a given source is owned by this socket.
